Add transitions to square element
When the mouse hovers over a square, the event mouseeenter is triggered,
which adds the hover class to the square element. Hover shortens the
transition time and changes the background color. When the mouse is moved
out of the square, the hover class is removed and the background color is
reverted back to the default transition time.
This creates the effect of quickly coloring the square the user is hovering
over, while fading away the user's path slowly, resulting in a trail.