Fix bug caused by Math.floor function
Previously, to determine the integer pixel width and height of each square,
the Math.floor function was used to round down the quotient from dividing
the total width of the grid by the number of squares per row/column.
With larger numbers of squares per row the rounding-down became significant,
causing more squares to appear in a row than intended. This fed bad data
into createPerimeterBorders(), resulting in a disfigured grid.

To correct for this, the Math.floor function was removed. While DevTools
displays an integer pixel value for each square, the grid appears to display
the correct number of squares per row.